+++ Project1/data/shaders/vertex.glsl
@@ -5,7 +5,7 @@ layout (location = 0) in vec3 aPos;
 layout (location = 1) in vec3 aColor;
 layout (location = 2) in vec2 aTexCoord;
 layout (location = 3) in vec3 aNormal;
-layout (location = 6) in uint aTextureID;
+layout (location = 6) in float aTextureID;
 
 #ifdef INSTANCED
 layout (location = 4) in mat4 model;
@@ -40,7 +40,7 @@ void main()
 	world = vec3(model * vec4(aPos, 1.0));
     normal = mat3(transpose(inverse(model))) * aNormal;  
     TexCoord = aTexCoord;
-	textureID = aTextureID;
+	textureID = uint(aTextureID);
 
 #ifdef NORMALMAP
 	computeTBN(mat3(transpose(inverse(model))), aNormal, tangent);
